Fix several bugs with undocumented feature of Bass.PluginLoad()
The first bug here is with path handling. The code splits the directory name from the file name then merges them together but does not put a file seperator between the two.

The second bug is with the state that it leaves Bass.LastError in after trying to load the plugin. If the plugin fails for any reason such as Errors.FileFormat, or Errors.Already that will get overwritten by later calls for the different platforms.